refactor: 시간 세는 부분 옮김
